+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Fixes
 
+* Suppress per-message notifications on Android and iOS for users with unknown usernames, and for one's own messages [#337](https://github.com/TryQuiet/quiet-private/issues/337)
 * iOS tor process lifecycle improvements solving crashes and improving performance [#3349](https://github.com/TryQuiet/quiet/issues/3349)
 * Update LFA to remove flaky timestamp validator [#3365](https://github.com/TryQuiet/quiet/issues/3365)
 * Improve image compression efficiency [#3364](https://github.com/TryQuiet/quiet/issues/3364)